+++ HISTORY.rst
@@ -182,6 +182,7 @@
 - Move the GitLab CI platform template into its own app named
   `platform_gitlab`.
 - Reduce the SSH connections used by the GitLab CI deploy stage.
+- Improve the method to pass custom arguments to Tesseract OCR.
 
 4.8.3 (2024-10-04)
 ==================